refactor: Reader의 Jdbc, Jpa 구현체와 인터페이스를 분리
