+++ src/main/java/roomescape/infrastructure/reader/JdbcUserReader.java
@@ -1,0 +1,41 @@
+package roomescape.infrastructure.reader;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.jdbc.core.simple.JdbcClient;
+import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
+import roomescape.business.application_service.reader.UserReader;
+import roomescape.business.dto.UserDto;
+import roomescape.exception.business.NotFoundException;
+
+import java.util.List;
+
+import static roomescape.exception.ErrorCode.USER_NOT_EXIST;
+
+@Repository
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class JdbcUserReader implements UserReader {
+
+    private final JdbcClient jdbcClient;
+
+    @Override
+    public List<UserDto> getAll() {
+        String sql = "SELECT * FROM users";
+
+        return jdbcClient.sql(sql)
+                .query(UserDto.ROW_MAPPER)
+                .list();
+    }
+
+    @Override
+    public UserDto getById(final String userIdValue) {
+        String sql = "SELECT * FROM users u WHERE u.id = :id";
+
+        return jdbcClient.sql(sql)
+                .param("id", userIdValue)
+                .query(UserDto.ROW_MAPPER)
+                .optional()
+                .orElseThrow(() -> new NotFoundException(USER_NOT_EXIST));
+    }
+}
