+++ src/main/java/roomescape/infrastructure/repository/JpaThemes.java
@@ -1,11 +1,11 @@
-package roomescape.infrastructure.jpa;
+package roomescape.infrastructure.repository;
 
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Repository;
 import roomescape.business.model.entity.Theme;
 import roomescape.business.model.repository.Themes;
 import roomescape.business.model.vo.Id;
-import roomescape.infrastructure.jpa.dao.JpaThemeDao;
+import roomescape.infrastructure.repository.dao.JpaThemeDao;
 
 import java.util.Optional;
