+++ src/main/java/roomescape/infrastructure/repository/dao/JpaReservationSlotDao.java
@@ -1,4 +1,4 @@
-package roomescape.infrastructure.jpa.dao;
+package roomescape.infrastructure.repository.dao;
 
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -12,7 +12,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface JpaReservationSlotDao extends JpaRepository<ReservationSlot, Id> {
-    
+
     @Query("""
             SELECT DISTINCT rs
             FROM ReservationSlot rs
