+++ src/test/java/roomescape/infrastructure/ReservationTimesTest.java
@@ -10,8 +10,8 @@ import org.springframework.context.annotation.Import;
 import roomescape.business.model.entity.ReservationTime;
 import roomescape.business.model.repository.ReservationTimes;
 import roomescape.business.model.vo.Id;
-import roomescape.infrastructure.jpa.JpaReservationSlots;
-import roomescape.infrastructure.jpa.JpaReservationTimes;
+import roomescape.infrastructure.repository.JpaReservationSlots;
+import roomescape.infrastructure.repository.JpaReservationTimes;
 import roomescape.test_util.JpaTestUtil;
 
 import java.time.LocalTime;
