+++ src/test/java/roomescape/infrastructure/ReservationsTest.java
@@ -16,11 +16,11 @@ import roomescape.business.model.repository.Reservations;
 import roomescape.business.model.repository.Themes;
 import roomescape.business.model.repository.Users;
 import roomescape.business.model.vo.Id;
-import roomescape.infrastructure.jpa.JpaReservationSlots;
-import roomescape.infrastructure.jpa.JpaReservationTimes;
-import roomescape.infrastructure.jpa.JpaReservations;
-import roomescape.infrastructure.jpa.JpaThemes;
-import roomescape.infrastructure.jpa.JpaUsers;
+import roomescape.infrastructure.repository.JpaReservationSlots;
+import roomescape.infrastructure.repository.JpaReservationTimes;
+import roomescape.infrastructure.repository.JpaReservations;
+import roomescape.infrastructure.repository.JpaThemes;
+import roomescape.infrastructure.repository.JpaUsers;
 import roomescape.test_util.JpaTestUtil;
 
 import java.time.LocalDate;
