+++ src/test/java/roomescape/infrastructure/UsersTest.java
@@ -9,7 +9,7 @@ import org.springframework.context.annotation.Import;
 import roomescape.business.model.entity.User;
 import roomescape.business.model.repository.Users;
 import roomescape.business.model.vo.Id;
-import roomescape.infrastructure.jpa.JpaUsers;
+import roomescape.infrastructure.repository.JpaUsers;
 import roomescape.test_util.JpaTestUtil;
 
 import java.util.Optional;
