+++ src/test/java/roomescape/test_util/JpaTestUtil.java
@@ -9,16 +9,16 @@ import roomescape.business.model.entity.ReservationTime;
 import roomescape.business.model.entity.Theme;
 import roomescape.business.model.entity.User;
 import roomescape.business.model.vo.Id;
-import roomescape.infrastructure.jpa.JpaReservationSlots;
-import roomescape.infrastructure.jpa.JpaReservationTimes;
-import roomescape.infrastructure.jpa.JpaReservations;
-import roomescape.infrastructure.jpa.JpaThemes;
-import roomescape.infrastructure.jpa.JpaUsers;
-import roomescape.infrastructure.jpa.dao.JpaReservationDao;
-import roomescape.infrastructure.jpa.dao.JpaReservationSlotDao;
-import roomescape.infrastructure.jpa.dao.JpaReservationTimeDao;
-import roomescape.infrastructure.jpa.dao.JpaThemeDao;
-import roomescape.infrastructure.jpa.dao.JpaUserDao;
+import roomescape.infrastructure.repository.JpaReservationSlots;
+import roomescape.infrastructure.repository.JpaReservationTimes;
+import roomescape.infrastructure.repository.JpaReservations;
+import roomescape.infrastructure.repository.JpaThemes;
+import roomescape.infrastructure.repository.JpaUsers;
+import roomescape.infrastructure.repository.dao.JpaReservationDao;
+import roomescape.infrastructure.repository.dao.JpaReservationSlotDao;
+import roomescape.infrastructure.repository.dao.JpaReservationTimeDao;
+import roomescape.infrastructure.repository.dao.JpaThemeDao;
+import roomescape.infrastructure.repository.dao.JpaUserDao;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
